[Java] Refactor to make min flow control strategy a special case of tagged.
